Create dedicated headers for memory config and layout enum
