+++ tests/tt_eager/tensors/test_copy_and_move.cpp
@@ -22,7 +22,6 @@
 #include "ttnn/cpp/ttnn/operations/experimental/reshape/view.hpp"
 #include "ttnn/decorators.hpp"
 #include "ttnn/operations/functions.hpp"
-#include "ttnn/tensor/enum_types.hpp"
 #include "ttnn/tensor/host_buffer/functions.hpp"
 #include "ttnn/tensor/shape/shape.hpp"
 #include "ttnn/tensor/storage.hpp"
