+++ tests/ttnn/unit_tests/gtests/tensor/test_tensor_layout.cpp
@@ -15,7 +15,6 @@
 #include "gtest/gtest.h"
 #include <tt-metalium/shape.hpp>
 #include "ttnn/operations/functions.hpp"
-#include "ttnn/tensor/enum_types.hpp"
 #include "ttnn/tensor/layout/alignment.hpp"
 #include "ttnn/tensor/layout/page_config.hpp"
 #include "ttnn/tensor/layout/tensor_layout.hpp"
