+++ tests/ttnn/unit_tests/gtests/tensor/test_xtensor_conversion.cpp
@@ -21,7 +21,6 @@
 #include <tt-metalium/shape.hpp>
 #include <tt_stl/span.hpp>
 #include "ttnn/operations/functions.hpp"
-#include "ttnn/tensor/enum_types.hpp"
 #include "ttnn/tensor/layout/tensor_layout.hpp"
 #include "ttnn/tensor/shape/shape.hpp"
 #include "ttnn/tensor/tensor.hpp"
