+++ tt-train/sources/ttml/core/ttnn_all_includes.hpp
@@ -72,8 +72,9 @@
 #include <ttnn/operations/moreh/moreh_sum/moreh_sum.hpp>                                                   // NOLINT
 #include <ttnn/operations/normalization/softmax/softmax.hpp>                                               // NOLINT
 #include <ttnn/operations/reduction/generic/generic_reductions.hpp>                                        // NOLINT
-#include <ttnn/tensor/enum_types.hpp>                                                                      // NOLINT
 #include <ttnn/tensor/host_buffer/functions.hpp>                                                           // NOLINT
+#include <ttnn/tensor/layout/layout.hpp>                                                                   // NOLINT
+#include <ttnn/tensor/memory_config/memory_config.hpp>                                                     // NOLINT
 #include <ttnn/tensor/tensor.hpp>                                                                          // NOLINT
 #include <ttnn/tensor/types.hpp>                                                                           // NOLINT
 #include <ttnn/tensor/xtensor/conversion_utils.hpp>                                                        // NOLINT
