+++ ttnn/api/ttnn/tensor/layout/tensor_layout.hpp
@@ -6,7 +6,6 @@
 
 #include <tt-metalium/shape2d.hpp>
 
-#include "ttnn/tensor/enum_types.hpp"
 #include "ttnn/tensor/layout/alignment.hpp"
 #include "ttnn/tensor/layout/page_config.hpp"
 #include "ttnn/tensor/shape/shape.hpp"
