+++ ttnn/cpp/ttnn/operations/experimental/auto_format/auto_format.hpp
@@ -5,7 +5,6 @@
 #pragma once
 #include "ttnn/tensor/tensor.hpp"
 #include "ttnn/tensor/types.hpp"
-#include "ttnn/tensor/enum_types.hpp"
 #include <tt-metalium/constants.hpp>
 #include "ttnn/operation.hpp"
