+++ ttnn/cpp/ttnn/operations/pool/upsample/device/upsample_op.cpp
@@ -5,7 +5,6 @@
 #include "upsample_op.hpp"
 
 #include "tt-metalium/buffer_types.hpp"
-#include "ttnn/tensor/enum_types.hpp"
 #include "ttnn/tensor/types.hpp"
 #include <tt-metalium/util.hpp>
 #include <tt-metalium/work_split.hpp>
